Terminado (en principio) AppMusic

+++ src/descuento/DescuentoJovenes.java
@@ -2,7 +2,7 @@ package descuento;
 
 public class DescuentoJovenes extends Descuento {
 	
-	private static final int PORCENTAJE_JOVENES = 20;
+	private static final float PORCENTAJE_JOVENES = (float)20.0;
 	
 	
 	public DescuentoJovenes()
@@ -13,6 +13,6 @@ public class DescuentoJovenes extends Descuento {
 	@Override
 	public float aplicarDescuento(float precio) 
 	{
-		return precio *(1 - PORCENTAJE_JOVENES / 100);
+		return precio *(1 - (float)(PORCENTAJE_JOVENES / 100.0));
 	}
 }
